moved isAprilFoolsDay out of SkyHanniDebugsAndTests.kt
